LayoutTests/imported/w3c/web-platform-tests/xhr/event-error.sub-expected.txt

Blocked access to external URL http://nonexistent-origin.localhost:8800/
CONSOLE MESSAGE: line 22: XMLHttpRequest cannot load http://nonexistent-origin.localhost:8800/ due to access control checks.

PASS XMLHttpRequest Test: event - error 
